+++ javaparser-core-testing/src/test/java/com/github/javaparser/ast/NodeTest.java
@@ -49,200 +49,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 class NodeTest {
-
-    @Test
-    void registerSubTree() {
-        String code = "class A { int f; void foo(int p) { return 'z'; }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-            @Override
-            public void propertyChange(Node observedNode, ObservableProperty property, Object oldValue, Object newValue) {
-                changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
-            }
-        };
-        cu.registerForSubtree(observer);
-
-        assertEquals(Arrays.asList(), changes);
-
-        cu.getClassByName("A").get().setName("MyCoolClass");
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass");
-
-        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean");
-
-        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean",
-                "Parameter.name changed from p to myParam");
-    }
-
-    @Test
-    void registerWithJustNodeMode() {
-        String code = "class A { int f; void foo(int p) { return 'z'; }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-            @Override
-            public void propertyChange(Node observedNode, ObservableProperty property, Object oldValue, Object newValue) {
-                changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
-            }
-        };
-        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.JUST_THIS_NODE);
-
-        assertEquals(Arrays.asList(), changes);
-
-        cu.getClassByName("A").get().setName("MyCoolClass");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
-
-        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
-
-        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
-
-        cu.getClassByName("MyCoolClass").get().addField("int", "bar").getVariables().get(0).setInitializer("0");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
-    }
-
-    @Test
-    void registerWithNodeAndExistingDescendantsMode() {
-        String code = "class A { int f; void foo(int p) { return 'z'; }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-            @Override
-            public void propertyChange(Node observedNode, ObservableProperty property, Object oldValue, Object newValue) {
-                changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
-            }
-        };
-        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.THIS_NODE_AND_EXISTING_DESCENDANTS);
-
-        assertEquals(Arrays.asList(), changes);
-
-        cu.getClassByName("A").get().setName("MyCoolClass");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
-
-        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean"), changes);
-
-        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean",
-                "Parameter.name changed from p to myParam"), changes);
-
-        cu.getClassByName("MyCoolClass").get().addField("int", "bar").getVariables().get(0).setInitializer("0");
-        assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean",
-                "Parameter.name changed from p to myParam"), changes);
-    }
-
-    @Test
-    void registerWithSelfPropagatingMode() {
-        String code = "class A { int f; void foo(int p) { return 'z'; }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-            @Override
-            public void propertyChange(Node observedNode, ObservableProperty property, Object oldValue, Object newValue) {
-                changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
-            }
-        };
-        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
-
-        assertThat(changes).isEmpty();
-
-        cu.getClassByName("A").get().setName("MyCoolClass");
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass");
-
-        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean");
-
-        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean",
-                "Parameter.name changed from p to myParam");
-
-        cu.getClassByName("MyCoolClass").get()
-                .addField("int", "bar")
-                .getVariables().get(0).setInitializer("0");
-        assertThat(changes).containsExactlyInAnyOrder("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
-                "FieldDeclaration.maximum_common_type changed from int to boolean",
-                "VariableDeclarator.type changed from int to boolean",
-                "Parameter.name changed from p to myParam",
-                "VariableDeclarator.initializer changed from null to 0");
-    }
-
-    @Test
-    void deleteAParameterTriggerNotifications() {
-        String code = "class A { void foo(int p) { }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-
-            @Override
-            public void listChange(NodeList observedNode, ListChangeType type, int index, Node nodeAddedOrRemoved) {
-                changes.add("removing [" + nodeAddedOrRemoved + "] from index " + index);
-            }
-        };
-        cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
-
-        cu.getClassByName("A").get().getMethodsByName("foo").get(0).getParameter(0).remove();
-        assertEquals(Arrays.asList("removing [int p] from index 0"), changes);
-    }
-
-    @Test
-    void deleteClassNameDoesNotTriggerNotifications() {
-        String code = "class A { void foo(int p) { }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-
-            @Override
-            public void listChange(NodeList observedNode, ListChangeType type, int index, Node nodeAddedOrRemoved) {
-                changes.add("removing [" + nodeAddedOrRemoved + "] from index " + index);
-            }
-        };
-        cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
-
-        // I cannot remove the name of a type
-        assertFalse(cu.getClassByName("A").get().getName().remove());
-        assertEquals(Arrays.asList(), changes);
-    }
-
-    @Test
-    void deleteMethodBodyDoesTriggerNotifications() {
-        String code = "class A { void foo(int p) { }}";
-        CompilationUnit cu = parse(code);
-        List<String> changes = new ArrayList<>();
-        AstObserver observer = new AstObserverAdapter() {
-
-            @Override
-            public void propertyChange(Node observedNode, ObservableProperty property, Object oldValue, Object newValue) {
-                changes.add("setting [" + property + "] to " + newValue);
-            }
-
-            @Override
-            public void listChange(NodeList observedNode, ListChangeType type, int index, Node nodeAddedOrRemoved) {
-                changes.add("removing [" + nodeAddedOrRemoved + "] from index " + index);
-            }
-        };
-        cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
-
-        assertTrue(cu.getClassByName("A").get().getMethodsByName("foo").get(0).getBody().get().remove());
-        assertEquals(Arrays.asList("setting [BODY] to null"), changes);
-    }
-
     @Test
     void removeOrphanCommentPositiveCase() {
         ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "A");
@@ -270,39 +76,34 @@ class NodeTest {
 
     @Test
     void hasJavaDocCommentPositiveCaseWithSetJavaDocComment() {
-        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(),
-                false, "Foo");
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "Foo");
         decl.setJavadocComment("A comment");
         assertTrue(decl.hasJavaDocComment());
     }
 
     @Test
     void hasJavaDocCommentPositiveCaseWithSetComment() {
-        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(),
-                false, "Foo");
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "Foo");
         decl.setComment(new JavadocComment("A comment"));
         assertTrue(decl.hasJavaDocComment());
     }
 
     @Test
     void hasJavaDocCommentNegativeCaseNoComment() {
-        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(),
-                false, "Foo");
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "Foo");
         assertFalse(decl.hasJavaDocComment());
     }
 
     @Test
     void hasJavaDocCommentNegativeCaseLineComment() {
-        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(),
-                false, "Foo");
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "Foo");
         decl.setComment(new LineComment("foo"));
         assertFalse(decl.hasJavaDocComment());
     }
 
     @Test
     void hasJavaDocCommentNegativeCaseBlockComment() {
-        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(),
-                false, "Foo");
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(new NodeList<>(), false, "Foo");
         decl.setComment(new BlockComment("foo"));
         assertFalse(decl.hasJavaDocComment());
     }
@@ -361,87 +162,5 @@ class NodeTest {
         target.remove();
         // This will throw an exception if the parents are bad.
         unit.toString();
-    }
-
-    @Test
-    void findCompilationUnit() {
-        CompilationUnit cu = parse("class X{int x;}");
-        VariableDeclarator x = cu.getClassByName("X").get().getMember(0).asFieldDeclaration().getVariables().get(0);
-        assertEquals(cu, x.findCompilationUnit().get());
-    }
-
-    @Test
-    void findParent() {
-        CompilationUnit cu = parse("class X{int x;}");
-        SimpleName x = cu.getClassByName("X").get().getMember(0).asFieldDeclaration().getVariables().get(0).getName();
-        assertEquals("int x;", x.findAncestor(FieldDeclaration.class).get().toString());
-    }
-
-    @Test
-    void cantFindCompilationUnit() {
-        VariableDeclarator x = new VariableDeclarator();
-        assertFalse(x.findCompilationUnit().isPresent());
-    }
-
-    @Test
-    void genericWalk() {
-        Expression e = parseExpression("1+1");
-        StringBuilder b = new StringBuilder();
-        e.walk(n -> b.append(n.toString()));
-        assertEquals("1 + 111", b.toString());
-    }
-
-    @Test
-    void classSpecificWalk() {
-        Expression e = parseExpression("1+1");
-        StringBuilder b = new StringBuilder();
-        e.walk(IntegerLiteralExpr.class, n -> b.append(n.toString()));
-        assertEquals("11", b.toString());
-    }
-
-    @Test
-    void conditionalFindAll() {
-        Expression e = parseExpression("1+2+3");
-        List<IntegerLiteralExpr> ints = e.findAll(IntegerLiteralExpr.class, n -> n.asInt() > 1);
-        assertEquals("[2, 3]", ints.toString());
-    }
-
-    @Test
-    void typeOnlyFindAll() {
-        Expression e = parseExpression("1+2+3");
-        List<IntegerLiteralExpr> ints = e.findAll(IntegerLiteralExpr.class);
-        assertEquals("[1, 2, 3]", ints.toString());
-    }
-
-    @Test
-    void typeOnlyFindAllMatchesSubclasses() {
-        Expression e = parseExpression("1+2+3");
-        List<Node> ints = e.findAll(Node.class);
-        assertEquals("[1 + 2 + 3, 1 + 2, 1, 2, 3]", ints.toString());
-    }
-
-    @Test
-    void conditionalTypedFindFirst() {
-        Expression e = parseExpression("1+2+3");
-        Optional<IntegerLiteralExpr> ints = e.findFirst(IntegerLiteralExpr.class, n -> n.asInt() > 1);
-        assertEquals("Optional[2]", ints.toString());
-    }
-
-    @Test
-    void typeOnlyFindFirst() {
-        Expression e = parseExpression("1+2+3");
-        Optional<IntegerLiteralExpr> ints = e.findFirst(IntegerLiteralExpr.class);
-        assertEquals("Optional[1]", ints.toString());
-    }
-
-    @Test
-    void stream() {
-        Expression e = parseExpression("1+2+3");
-        List<IntegerLiteralExpr> ints = e.stream()
-                .filter(n -> n instanceof IntegerLiteralExpr)
-                .map(IntegerLiteralExpr.class::cast)
-                .filter(i -> i.asInt() > 1)
-                .collect(Collectors.toList());
-        assertEquals("[2, 3]", ints.toString());
     }
 }
