+++ javaparser-core/src/main/java/com/github/javaparser/ast/observer/AstObserverAdapter.java
@@ -37,12 +37,12 @@ public abstract class AstObserverAdapter implements AstObserver {
     }
 
     @Override
-    public void listChange(NodeList observedNode, ListChangeType type, int index, Node nodeAddedOrRemoved) {
+    public void listChange(NodeList<?> observedNode, ListChangeType type, int index, Node nodeAddedOrRemoved) {
         // do nothing
     }
 
     @Override
-    public void listReplacement(NodeList observedNode, int index, Node oldNode, Node newNode) {
+    public void listReplacement(NodeList<?> observedNode, int index, Node oldNode, Node newNode) {
         // do nothing
     }
 }
